delete old use-drop-target, fix findClosestDrop...
