+++ packages/design-system/src/components/primitives/dnd/index.ts
@@ -1,4 +1,4 @@
 export * from "./placement";
 export * from "./use-drag";
-export * from "./use-drop-target2";
+export * from "./use-drop-target";
 export * from "./use-auto-scroll";
